+++ db/migrate/20230606182219_add_full_name_to_hiring_agreements_signatures.rb
@@ -1,0 +1,6 @@
+class AddFullNameToHiringAgreementsSignatures < ActiveRecord::Migration[7.0]
+  def change
+    add_column :hiring_agreements_signatures, :full_name, :string
+    add_column :hiring_agreements_signatures, :ip_address, :string
+  end
+end